pyproject.toml
[project]
name = "ytdlbot"
version = "0.2.0"
description = "Telegram Video Downloader Bot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.7",
    "urllib3>=2.3.0",
    "validators>=0.34.0",
    "yt-dlp>=2024.11.4",
]
